requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Mail
Flask-Login
requests
gunicorn
pymysql